Implement RabbitMQ survey action
Matched patterns can be published as message to RMQ.
Each pattern can define sucess action as rmq_survey if it wants
to be published to RMQ.
The details of RMQ has to be configured under config block [rmq_survey].
The implementation is pushing messages to RMQ on exchange which is of
type topic. Also the message are to be forwarded to a specific configured
queue and not just exchange.
Currently using Blocking adapter to send msgs to RMQ broker.