Handle control characters in LogSurvey
Log format from config file can pass control characters like \n
or \r and these should be interpreted, substituted and not written
to log file as is.
At the same time, control characters read from target file should
not be substituted and should be writted as is; as control characters
like \r are malforming the message when they are logged.